+++ tests/utils.py
@@ -718,17 +718,23 @@ def mocked_execute_process():
     )
 
 
+@contextlib.contextmanager
 def mocked_dismiss_process():
     """
     Mock operations called to terminate :mod:`Celery` tasks.
+
+    Can be used either as decorator or context.
     """
     mock_celery_app = mock.MagicMock()
     mock_celery_app.control = mock.MagicMock()
     mock_celery_app.control.revoke = mock.MagicMock()
+    mock_celery_revoke = mock.patch("weaver.wps_restapi.jobs.jobs.celery_app", return_value=mock_celery_app)
 
-    return (
-        mock.patch("weaver.wps_restapi.jobs.jobs.app", return_value=mock_celery_app)
-    )
+    try:
+        with mock_celery_revoke:
+            yield   # for direct use by context or decorator
+    finally:
+        return mock_celery_revoke  # for use by combined ExitStack context  # pylint: disable=W0150.lost-exception
 
 
 def mocked_process_job_runner(job_task_id="mocked-job-id"):