+++ tests/wps_restapi/test_jobs.py
@@ -1,6 +1,9 @@
 import contextlib
 import datetime
 import json
+import os
+import shutil
+import tempfile
 import unittest
 import warnings
 from collections import OrderedDict
@@ -16,6 +19,7 @@ from dateutil import parser as date_parser
 from tests.utils import (
     get_module_version,
     get_test_weaver_app,
+    mocked_dismiss_process,
     mocked_process_job_runner,
     mocked_remote_wps,
     setup_config_with_mongodb,
@@ -32,8 +36,10 @@ from weaver.status import (
     JOB_STATUS_CATEGORY_FINISHED,
     JOB_STATUS_VALUES,
     STATUS_ACCEPTED,
+    STATUS_DISMISSED,
     STATUS_FAILED,
     STATUS_RUNNING,
+    STATUS_STARTED,
     STATUS_SUCCEEDED
 )
 from weaver.utils import get_path_kvp, now
@@ -60,6 +66,7 @@ class WpsRestApiJobsTest(unittest.TestCase):
         cls.settings = {
             "weaver.url": "https://localhost",
             "weaver.wps_email_encrypt_salt": "weaver-test",
+            "weaver.wps_output_dir": "/tmp/weaver-test/wps-outputs",
         }
         cls.config = setup_config_with_mongodb(settings=cls.settings)
         cls.app = get_test_weaver_app(config=cls.config)
@@ -106,43 +113,49 @@ class WpsRestApiJobsTest(unittest.TestCase):
         self.make_job(task_id="0000-0000-0000-0000",
                       process=self.process_public.identifier, service=None,
                       user_id=self.user_editor1_id, status=STATUS_SUCCEEDED, progress=100, access=VISIBILITY_PUBLIC)
-        self.make_job(task_id="1111-1111-1111-1111",
+        self.make_job(task_id="0000-0000-0000-1111",
                       process=self.process_unknown, service=self.service_public.name,
                       user_id=self.user_editor1_id, status=STATUS_FAILED, progress=99, access=VISIBILITY_PUBLIC)
-        self.make_job(task_id="2222-2222-2222-2222",
+        self.make_job(task_id="0000-0000-0000-2222",
                       process=self.process_private.identifier, service=None,
                       user_id=self.user_editor1_id, status=STATUS_FAILED, progress=55, access=VISIBILITY_PUBLIC)
         # same process as job 0, but private (ex: job ran with private process, then process made public afterwards)
-        self.make_job(task_id="3333-3333-3333-3333",
+        self.make_job(task_id="0000-0000-0000-3333",
                       process=self.process_public.identifier, service=None,
                       user_id=self.user_editor1_id, status=STATUS_FAILED, progress=55, access=VISIBILITY_PRIVATE)
         # job ran by admin
-        self.make_job(task_id="4444-4444-4444-4444",
+        self.make_job(task_id="0000-0000-0000-4444",
                       process=self.process_public.identifier, service=None,
                       user_id=self.user_admin_id, status=STATUS_FAILED, progress=55, access=VISIBILITY_PRIVATE)
         # job public/private service/process combinations
-        self.make_job(task_id="5555-5555-5555-5555", created=self.datetime_interval[0], duration=20,
+        self.make_job(task_id="0000-0000-0000-5555", created=self.datetime_interval[0], duration=20,
                       process=self.process_public.identifier, service=self.service_public.name,
                       user_id=self.user_editor1_id, status=STATUS_FAILED, progress=99, access=VISIBILITY_PUBLIC)
-        self.make_job(task_id="6666-6666-6666-6666", created=self.datetime_interval[1], duration=30,
+        self.make_job(task_id="0000-0000-0000-6666", created=self.datetime_interval[1], duration=30,
                       process=self.process_private.identifier, service=self.service_public.name,
                       user_id=self.user_editor1_id, status=STATUS_FAILED, progress=99, access=VISIBILITY_PUBLIC)
-        self.make_job(task_id="7777-7777-7777-7777", created=self.datetime_interval[2], duration=40,
+        self.make_job(task_id="0000-0000-0000-7777", created=self.datetime_interval[2], duration=40,
                       process=self.process_public.identifier, service=self.service_private.name,
                       user_id=self.user_editor1_id, status=STATUS_FAILED, progress=99, access=VISIBILITY_PUBLIC)
-        self.make_job(task_id="8888-8888-8888-8888", created=self.datetime_interval[3], duration=50,
+        self.make_job(task_id="0000-0000-0000-8888", created=self.datetime_interval[3], duration=50,
                       process=self.process_private.identifier, service=self.service_private.name,
                       user_id=self.user_editor1_id, status=STATUS_FAILED, progress=99, access=VISIBILITY_PUBLIC)
         # jobs with duplicate 'process' identifier, but under a different 'service' name
         # WARNING:
         #   For tests that use minDuration/maxDuration, following two jobs could 'eventually' become more/less than
         #   expected test values while debugging (code breakpoints) since their duration is dynamic (current - started)
-        self.make_job(task_id="9999-9999-9999-9999", created=now(), duration=20,
+        self.make_job(task_id="0000-0000-0000-9999", created=now(), duration=20,
                       process=self.process_other.identifier, service=self.service_one.name,
                       user_id=self.user_editor1_id, status=STATUS_RUNNING, progress=99, access=VISIBILITY_PUBLIC)
-        self.make_job(task_id="1010-1010-1010-1010", created=now(), duration=25,
+        self.make_job(task_id="0000-0000-1111-0000", created=now(), duration=25,
                       process=self.process_other.identifier, service=self.service_two.name,
                       user_id=self.user_editor1_id, status=STATUS_RUNNING, progress=99, access=VISIBILITY_PUBLIC)
+        self.make_job(task_id="0000-0000-2222-0000", created=now(), duration=0,
+                      process=self.process_other.identifier, service=self.service_two.name,
+                      user_id=self.user_editor1_id, status=STATUS_ACCEPTED, progress=99, access=VISIBILITY_PUBLIC)
+        self.make_job(task_id="0000-0000-3333-0000", created=now(), duration=0,
+                      process=self.process_other.identifier, service=self.service_two.name,
+                      user_id=self.user_editor1_id, status=STATUS_STARTED, progress=99, access=VISIBILITY_PUBLIC)
 
     def make_job(self, task_id, process, service, user_id, status, progress, access,
                  created=None, offset=None, duration=None):
@@ -169,14 +182,23 @@ class WpsRestApiJobsTest(unittest.TestCase):
         return message + "\nMapping Task-ID/Job-ID:\n{}".format(json.dumps(mapping, indent=indent))
 
     def assert_equal_with_jobs_diffs(self, jobs_result, jobs_expect,
-                                     test_values=None, message="", indent=2, index=None):
-        assert len(jobs_result) == len(jobs_expect) and all(job in jobs_expect for job in jobs_result), \
-            (message if message else "Different jobs returned than expected") + \
-            (" (index: {})".format(index) if index is not None else "") + \
-            ("\nResponse: {}".format(json.dumps(sorted(jobs_result), indent=indent))) + \
-            ("\nExpected: {}".format(json.dumps(sorted(jobs_expect), indent=indent))) + \
-            ("\nTesting: {}".format(test_values) if test_values else "") + \
+                                     test_values=None, message="", indent=2, index=None, invert=False):
+        mapping = {job.id: job.task_id for job in self.job_info}
+        missing = set(jobs_expect) - set(jobs_result)
+        unknown = set(jobs_result) - set(jobs_expect)
+        assert (
+            (invert or len(jobs_result) == len(jobs_expect)) and
+            all((job not in jobs_expect if invert else job in jobs_expect) for job in jobs_result)
+        ), (
+            (message if message else "Different jobs returned than expected") +
+            (" (index: {})".format(index) if index is not None else "") +
+            ("\nResponse: {}".format(json.dumps(sorted(jobs_result), indent=indent))) +
+            ("\nExpected: {}".format(json.dumps(sorted(jobs_expect), indent=indent))) +
+            ("\nMissing: {}".format(json.dumps(sorted(f"{job} ({mapping[job]})" for job in missing), indent=indent))) +
+            ("\nUnknown: {}".format(json.dumps(sorted(f"{job} ({mapping[job]})" for job in unknown), indent=indent))) +
+            ("\nTesting: {}".format(test_values) if test_values else "") +
             (self.message_with_jobs_mapping())
+        )
 
     def get_job_request_auth_mock(self, user_id):
         is_admin = self.user_admin_id == user_id
@@ -316,8 +338,8 @@ class WpsRestApiJobsTest(unittest.TestCase):
                 assert len(grouped_jobs["jobs"]) == 1
                 assert set(grouped_jobs["jobs"]) == {self.job_info[1].id}
             elif categories["process"] == self.process_other.identifier:
-                assert len(grouped_jobs["jobs"]) == 2
-                assert set(grouped_jobs["jobs"]) == {self.job_info[9].id, self.job_info[10].id}
+                assert len(grouped_jobs["jobs"]) == 4
+                assert set(grouped_jobs["jobs"]) == {self.job_info[i].id for i in [9, 10, 11, 12]}
             else:
                 pytest.fail("Unknown job grouping 'process' value: {}".format(categories["process"]))
 
@@ -346,8 +368,8 @@ class WpsRestApiJobsTest(unittest.TestCase):
                 assert len(grouped_jobs["jobs"]) == 1
                 assert set(grouped_jobs["jobs"]) == {self.job_info[9].id}
             elif categories[service_or_provider] == self.service_two.name:
-                assert len(grouped_jobs["jobs"]) == 1
-                assert set(grouped_jobs["jobs"]) == {self.job_info[10].id}
+                assert len(grouped_jobs["jobs"]) == 3
+                assert set(grouped_jobs["jobs"]) == {self.job_info[10].id, self.job_info[11].id, self.job_info[12].id}
             elif categories[service_or_provider] is None:
                 assert len(grouped_jobs["jobs"]) == 2
                 assert set(grouped_jobs["jobs"]) == {self.job_info[0].id, self.job_info[2].id}
@@ -386,14 +408,15 @@ class WpsRestApiJobsTest(unittest.TestCase):
 
         base_url = self.settings["weaver.url"]
         jobs_url = base_url + sd.jobs_service.path
-        limit = 2  # expect 9 jobs to be visible, making 5 pages of 2
-        last = 4
+        limit = 2  # expect 11 jobs to be visible, making 6 pages of 2 each (except last that is 1)
+        last = 5   # zero-based index of last page
         last_page = "page={}".format(last)
         prev_last_page = "page={}".format(last - 1)
         limit_kvp = "limit={}".format(limit)
         path = get_path_kvp(sd.jobs_service.path, limit=limit)
         resp = self.app.get(path, headers=self.json_headers)
         links = get_links(resp.json["links"])
+        assert resp.json["total"] == expect_jobs_visible
         assert len(resp.json["jobs"]) == limit
         assert links["alternate"] is None
         assert links["collection"] == jobs_url
@@ -473,9 +496,9 @@ class WpsRestApiJobsTest(unittest.TestCase):
         assert links["last"].startswith(p_j_url) and limit_kvp in links["last"] and "page=1" in links["last"]
         assert all(p_kvp not in links[rel] for rel in ["current", "next", "prev", "first", "last"] if links[rel])
 
-        over_limit = 10
-        limit_kvp = "limit={}".format(over_limit)
-        path = get_path_kvp(sd.jobs_service.path, limit=over_limit)
+        limit_over_total = expect_jobs_visible * 2
+        limit_kvp = "limit={}".format(limit_over_total)
+        path = get_path_kvp(sd.jobs_service.path, limit=limit_over_total)
         resp = self.app.get(path, headers=self.json_headers)
         links = get_links(resp.json["links"])
         assert len(resp.json["jobs"]) == expect_jobs_visible
@@ -596,7 +619,7 @@ class WpsRestApiJobsTest(unittest.TestCase):
         path = get_path_kvp(sd.jobs_service.path, type="provider", process=self.process_other.identifier, detail=True)
         resp = self.app.get(path, headers=self.json_headers)
         self.check_basic_jobs_info(resp)
-        expect_jobs = [self.job_info[i].id for i in [9, 10]]
+        expect_jobs = [self.job_info[i].id for i in [9, 10, 11, 12]]
         result_jobs = [job["jobID"] for job in resp.json["jobs"]]
         self.assert_equal_with_jobs_diffs(result_jobs, expect_jobs)
         assert resp.json["total"] == len(expect_jobs)
@@ -839,15 +862,17 @@ class WpsRestApiJobsTest(unittest.TestCase):
         assert resp.content_type == CONTENT_TYPE_APP_JSON
         # generated datetime interval have an offset that makes all job in the future
         # anything created "recently" and publicly visible will be listed here
-        assert len(resp.json["jobs"]) == 6
+        job_result = resp.json["jobs"]
+        job_expect = [self.job_info[i].id for i in [0, 1, 2, 5, 9, 10, 11, 12]]
+        self.assert_equal_with_jobs_diffs(job_result, job_expect, {"datetime": datetime_before})
         for job in resp.json["jobs"]:
             base_uri = sd.jobs_service.path + "/{}".format(job)
             path = get_path_kvp(base_uri)
             resp = self.app.get(path, headers=self.json_headers)
             assert resp.status_code == 200
             assert resp.content_type == CONTENT_TYPE_APP_JSON
-            assert date_parser.parse(resp.json["created"]) <= date_parser.parse(
-                datetime_before.replace(DATETIME_INTERVAL_OPEN_START_SYMBOL, ""))
+            interval = datetime_before.replace(DATETIME_INTERVAL_OPEN_START_SYMBOL, "")
+            assert date_parser.parse(resp.json["created"]) <= date_parser.parse(interval)
 
     def test_get_jobs_datetime_after(self):
         """
@@ -869,8 +894,8 @@ class WpsRestApiJobsTest(unittest.TestCase):
             resp = self.app.get(path, headers=self.json_headers)
             assert resp.status_code == 200
             assert resp.content_type == CONTENT_TYPE_APP_JSON
-            assert date_parser.parse(resp.json["created"]) >= date_parser.parse(
-                datetime_after.replace(DATETIME_INTERVAL_OPEN_END_SYMBOL, ""))
+            interval = datetime_after.replace(DATETIME_INTERVAL_OPEN_END_SYMBOL, "")
+            assert date_parser.parse(resp.json["created"]) >= date_parser.parse(interval)
 
     def test_get_jobs_datetime_interval(self):
         """
@@ -998,7 +1023,8 @@ class WpsRestApiJobsTest(unittest.TestCase):
         resp = self.app.get(path, headers=self.json_headers)
         assert resp.status_code == 200
         result_jobs = resp.json["jobs"]
-        expect_idx = [0, 1, 2, 5, 6, 9, 10]  # 3, 4 are private, 9, 10 dynamic since running, others fixed < 30
+        # 3, 4 are private, >9 except 11 are dynamic since running (11 only accepted), others fixed duration <30s
+        expect_idx = [0, 1, 2, 5, 6, 9, 10, 12]
         expect_jobs = [self.job_info[i].id for i in expect_idx]
         self.assert_equal_with_jobs_diffs(result_jobs, expect_jobs, test)
 
@@ -1007,7 +1033,8 @@ class WpsRestApiJobsTest(unittest.TestCase):
         resp = self.app.get(path, headers=self.json_headers)
         assert resp.status_code == 200
         result_jobs = resp.json["jobs"]
-        expect_idx = [0, 1, 2, 5, 6, 7, 9, 10]  # everything except 3, 4 that are private, and 8 that is == 50
+        # same as previous for repeated indices except 7 == 40s now also < max duration, 8 is 50s just below range
+        expect_idx = [0, 1, 2, 5, 6, 7, 9, 10, 12]
         expect_jobs = [self.job_info[i].id for i in expect_idx]
         self.assert_equal_with_jobs_diffs(result_jobs, expect_jobs, test)
 
@@ -1118,3 +1145,122 @@ class WpsRestApiJobsTest(unittest.TestCase):
             assert resp.content_type == CONTENT_TYPE_APP_JSON
 
         assert resp.json["processID"] == "process-public"
+
+    @mocked_dismiss_process()
+    def test_job_dismiss_running_single(self):
+        """
+        Jobs that are in a valid *running* (or about to) state can be dismissed successfully.
+
+        Subsequent calls to the same job dismiss operation must respond with HTTP Gone (410) status.
+
+        .. seealso::
+            OGC specification of dismiss operation: https://docs.ogc.org/DRAFTS/18-062.html#sec_cons_dismiss
+        """
+        job_running = self.job_info[10]
+        assert job_running.status == STATUS_RUNNING, "Job must be in running state for test"
+        job_accept = self.job_info[11]
+        assert job_accept.status == STATUS_ACCEPTED, "Job must be in accepted state for test"
+        job_started = self.job_info[12]
+        assert job_started.status == STATUS_STARTED, "Job must be in started state for test"
+
+        for job in [job_running, job_accept, job_started]:
+            path = sd.job_service.path.format(job_id=job.id)
+            resp = self.app.delete(path, headers=self.json_headers)
+            assert resp.status_code == 200
+            assert resp.json["status"] == STATUS_DISMISSED
+
+            # job are not removed, only dismissed
+            path = get_path_kvp(sd.jobs_service.path, status=STATUS_DISMISSED, limitt=1000)
+            resp = self.app.get(path, headers=self.json_headers)
+            assert resp.status_code == 200
+            assert job.id in resp.json["jobs"], "Job HTTP DELETE should not have deleted it, but only dismissed it."
+
+            path = sd.job_service.path.format(job_id=job.id)
+            resp = self.app.get(path, headers=self.json_headers)
+            assert resp.status_code == 200, "Job should still exist even after dismiss"
+            assert resp.json["status"] == STATUS_DISMISSED
+
+            resp = self.app.delete(path, headers=self.json_headers, expect_errors=True)
+            assert resp.status_code == 410, "Job cannot be dismissed again."
+            assert job.id in resp.json["value"]
+
+    @mocked_dismiss_process()
+    def test_job_dismiss_complete_single(self):
+        """
+        Jobs that are already *completed* (regardless of success/failure) state removes result artifacts.
+
+        Subsequent calls to the same job dismiss operation must respond with HTTP Gone (410) status.
+
+        .. seealso::
+            OGC specification of dismiss operation: https://docs.ogc.org/DRAFTS/18-062.html#sec_cons_dismiss
+        """
+        job_success = self.job_info[0]
+        job_failed = self.job_info[1]
+        assert job_success.status == STATUS_SUCCEEDED, "Job must be in successful state for test"
+        assert job_failed.status == STATUS_FAILED, "Job must be in failed state for test"
+
+        # create dummy files to validate results flush of successful job
+        wps_out_dir = self.settings["weaver.wps_output_dir"]
+        job_out_dir = os.path.join(wps_out_dir, job_success.id)
+        job_out_log = os.path.join(wps_out_dir, job_success.id + ".log")
+        job_out_xml = os.path.join(wps_out_dir, job_success.id + ".xml")
+        os.makedirs(job_out_dir, exist_ok=True)
+        try:
+            with contextlib.ExitStack() as stack:
+                tmp_out1 = stack.enter_context(tempfile.NamedTemporaryFile(mode="w", dir=job_out_dir, suffix=".yml"))
+                tmp_out2 = stack.enter_context(tempfile.NamedTemporaryFile(mode="w", dir=job_out_dir, suffix=".txt"))
+                tmp_out3 = stack.enter_context(tempfile.NamedTemporaryFile(mode="w", dir=job_out_dir, suffix=".tif"))
+                tmp_log = stack.enter_context(open(job_out_log, "w"))  # noqa
+                tmp_xml = stack.enter_context(open(job_out_xml, "w"))  # noqa
+                for tmp_file in [tmp_out1, tmp_out2, tmp_out3, tmp_log, tmp_xml]:
+                    assert os.path.isfile(tmp_file.name)
+
+                job_path = sd.job_service.path.format(job_id=job_success.id)
+                resp = self.app.delete(job_path, headers=self.json_headers)
+                assert resp.status_code == 200
+                assert resp.json["status"] == STATUS_DISMISSED
+
+                for tmp_file in [tmp_out1, tmp_out2, tmp_out3, tmp_log, tmp_xml]:
+                    assert not os.path.exists(tmp_file.name)
+                assert not os.path.exists(job_out_dir)
+
+                # subsequent operations returns Gone for sub-resources of the job execution
+                for sub_path in ["", "/outputs", "/results", "/logs", "/exceptions"]:
+                    path = job_path + sub_path
+                    func = self.app.get if sub_path else self.app.delete
+                    resp = func(path, headers=self.json_headers, expect_errors=True)
+                    assert resp.status_code == 410, "Dismissed job should return 'Gone' status for: [{}]".format(path)
+        except OSError:
+            pass
+        finally:
+            shutil.rmtree(job_out_dir, ignore_errors=True)
+
+        # test on failed job that could have no artifacts at all
+        path = sd.job_service.path.format(job_id=job_failed.id)
+        resp = self.app.delete(path, headers=self.json_headers)
+        assert resp.status_code == 200
+        assert resp.json["status"] == STATUS_DISMISSED
+        resp = self.app.delete(path, headers=self.json_headers, expect_errors=True)
+        assert resp.status_code == 410
+
+    @mocked_dismiss_process()
+    def test_job_dismiss_batch(self):
+        path = sd.jobs_service.path
+        resp = self.app.get(path, headers=self.json_headers)
+        assert resp.status_code == 200
+        jobs = resp.json["jobs"]
+        assert len(jobs) > 3
+
+        resp = self.app.delete_json(path, params={"jobs": jobs[:2]}, headers=self.json_headers)
+        assert resp.status_code == 200
+        self.assert_equal_with_jobs_diffs(resp.json["jobs"], jobs[:2])
+
+        resp = self.app.get(path, headers=self.json_headers)
+        assert resp.status_code == 200
+        self.assert_equal_with_jobs_diffs(resp.json["jobs"], jobs, message="All jobs should still exist after dismiss.")
+
+        for job in jobs[:2]:
+            path = sd.job_service.path.format(job_id=job)
+            resp = self.app.get(path, headers=self.json_headers)
+            assert resp.status_code == 200
+            assert resp.json["status"] == STATUS_DISMISSED, "Job status should have been updated to dismissed."