+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changes
 
 .. **REPLACE AND/OR ADD SECTION ENTRIES ACCORDINGLY WITH APPLIED CHANGES**
 
-TMP FUTURE RELEASE
-==================
+`Unreleased <https://github.com/crim-ca/weaver/tree/master>`_ (latest)
+========================================================================
 
 Changes:
 --------
@@ -28,7 +28,7 @@ Fixes:
 - Fix incorrect resolution of ``Process`` results endpoint to pass contents from one step to another
   during ``Workflow`` execution (resolves `#358 <https://github.com/crim-ca/weaver/issues/358>`_).
 
-`Unreleased <https://github.com/crim-ca/weaver/tree/master>`_ (latest)
+`4.3.0 <https://github.com/crim-ca/weaver/tree/4.3.0>`_ (2021-11-16)
 ========================================================================
 
 Changes:
@@ -59,6 +59,9 @@ Changes:
 - Avoid ``Job.progress`` updates following ``failed`` or ``dismissed`` statuses to keep track of the last real progress
   percentage that was reached when that status was set.
 - Improve typing of database and store getter functions to infer correct types and facilitate code auto-complete.
+- Implement ``Job`` `dismiss operation <https://docs.ogc.org/DRAFTS/18-062.html#sec_cons_dismiss>`_ ensuring
+  pending or running tasks are removed and output result artifacts are removed from disk.
+- Implement HTTP Gone (410) status from already dismissed ``Job`` when requested again or when fetching its artifacts.
 
 Fixes:
 ------
@@ -68,6 +71,10 @@ Fixes:
   instead of `Pyramid` ``Request``, as employed by underlying HTTP requests in `PyWPS` service.
 - Allow ``group`` query parameter to handle ``Job`` category listing with ``provider`` as ``service`` alias.
 - Improve typing of database and store getter functions to infer correct types and facilitate code auto-complete.
+- Fix incorrectly configured API views for batch ``Job`` dismiss operation with ``DELETE /jobs`` and corresponding
+  endpoints for ``Process`` and ``Provider`` paths.
+- Fix invalid ``Job`` links sometimes containing duplicate ``/`` occurrences.
+- Fix invalid ``Job`` link URL for ``alternate`` relationship.
 
 `4.2.1 <https://github.com/crim-ca/weaver/tree/4.2.1>`_ (2021-10-20)
 ========================================================================